Fix being loaded by Scalar::Util
Older versions of Scalar::Util that rely on List::Util for their
implementation load List::Util before declaring a version.  This means
we don't know what version to provide to XSLoader.  However, as long as
there is no $VERSION in existance, we can just not provide it a version
number and it will accept it.  We are already removing the VERSION by
clearing out the stash.  Also, if $Scalar::Util::VERSION is not defined
after requiring the module, we know that it is an old version loading us
mid way, and can proceed accordingly.