Try to use XS implementation if possible and fall back to PP
If the XS module is available, use it to implement the functions.  If it
isn't available, use the included PP module.

Older versions of List::Util shipped with an XS module that did nothing
but load List::Util itself.  We need to make sure we have a new enough
version that provides its own implementation.